feat: :art: Divide modal & edit form from profile page
