+++ frontend/src/pages/ProfilePage/EditForm.tsx
@@ -1,0 +1,7 @@
+export const EditForm = () => {
+  return (
+    <div>
+      <p>Modal form text</p>
+    </div>
+  );
+};
